# backup/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DevCamper API",
    "version": "1.0.0",
    "description": "A comprehensive API for managing bootcamps, courses, reviews, and users.",
    "contact": {
      "name": "API Support",
      "email": "support@devcamper.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4444/api/v1",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Bootcamp": {
        "type": "object",
        "required": [
          "name",
          "description",
          "address"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The auto-generated id of the bootcamp"
          },
          "name": {
            "type": "string",
            "description": "The name of the bootcamp"
          },
          "description": {
            "type": "string",
            "description": "The description of the bootcamp"
          },
          "website": {
            "type": "string",
            "description": "The website of the bootcamp"
          },
          "phone": {
            "type": "string",
            "description": "The phone number of the bootcamp"
          },
          "email": {
            "type": "string",
            "description": "The email of the bootcamp"
          },
          "address": {
            "type": "string",
            "description": "The address of the bootcamp"
          },
          "careers": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The careers offered by the bootcamp"
          },
          "housing": {
            "type": "boolean",
            "description": "Whether housing is provided"
          },
          "jobAssistance": {
            "type": "boolean",
            "description": "Whether job assistance is provided"
          },
          "jobGuarantee": {
            "type": "boolean",
            "description": "Whether job guarantee is provided"
          },
          "acceptGi": {
            "type": "boolean",
            "description": "Whether GI Bill is accepted"
          }
        },
        "example": {
          "name": "Devworks Bootcamp",
          "description": "This is a bootcamp for developers",
          "website": "https://devworks.com",
          "phone": "111-111-1111",
          "email": "user@gmail.com",
          "address": "23 Main St, Boston MA",
          "careers": [
            "Web Development",
            "UI/UX",
            "Business"
          ],
          "housing": true,
          "jobAssistance": true,
          "jobGuarantee": false,
          "acceptGi": true
        }
      },
      "Course": {
        "type": "object",
        "required": [
          "title",
          "description",
          "weeks",
          "tuition",
          "minimumSkill",
          "bootcamp"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The auto-generated id of the course"
          },
          "title": {
            "type": "string",
            "description": "The title of the course"
          },
          "description": {
            "type": "string",
            "description": "The description of the course"
          },
          "weeks": {
            "type": "string",
            "description": "The duration of the course in weeks"
          },
          "tuition": {
            "type": "number",
            "description": "The tuition cost of the course"
          },
          "minimumSkill": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ],
            "description": "The minimum skill required for the course"
          },
          "scholarshipsAvailable": {
            "type": "boolean",
            "description": "Whether scholarships are available"
          },
          "bootcamp": {
            "type": "string",
            "description": "The ID of the bootcamp the course belongs to"
          }
        },
        "example": {
          "title": "Front End Web Development",
          "description": "This course will teach you how to build high quality web applications with technologies like React, Redux, Node.js, PHP & Laravel",
          "weeks": 8,
          "tuition": 8000,
          "minimumSkill": "intermediate",
          "scholarshipsAvailable": true,
          "bootcamp": "5d713995b721c3bb38c1f5d0"
        }
      },
      "Review": {
        "type": "object",
        "required": [
          "title",
          "text",
          "rating",
          "bootcamp",
          "user"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The auto-generated id of the review"
          },
          "title": {
            "type": "string",
            "description": "The title of the review"
          },
          "text": {
            "type": "string",
            "description": "The text of the review"
          },
          "rating": {
            "type": "number",
            "description": "The rating of the review"
          },
          "bootcamp": {
            "type": "string",
            "description": "The ID of the bootcamp the review belongs to"
          },
          "user": {
            "type": "string",
            "description": "The ID of the user who wrote the review"
          }
        },
        "example": {
          "title": "Great Bootcamp",
          "text": "This bootcamp was amazing!",
          "rating": 9,
          "bootcamp": "5d713995b721c3bb38c1f5d0",
          "user": "5d7a514b5d2c12c7449be045"
        }
      },
      "User": {
        "type": "object",
        "required": [
          "name",
          "email",
          "password"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The auto-generated id of the user"
          },
          "name": {
            "type": "string",
            "description": "The name of the user"
          },
          "email": {
            "type": "string",
            "description": "The email of the user"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "publisher"
            ],
            "description": "The role of the user"
          },
          "password": {
            "type": "string",
            "description": "The password of the user"
          },
          "createdAt": {
            "type": "string",
            "format": "date",
            "description": "The date the user was created"
          }
        },
        "example": {
          "name": "John Doe",
          "email": "john@gmail.com",
          "role": "user",
          "password": "password123"
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/bootcamps": {
      "get": {
        "summary": "Get all bootcamps",
        "tags": [
          "Bootcamps"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "pagination": {
                      "type": "object"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Bootcamp"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new bootcamp",
        "tags": [
          "Bootcamps"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Bootcamp"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/courses": {
      "get": {
        "summary": "Get all courses",
        "tags": [
          "Courses"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "pagination": {
                      "type": "object"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Course"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new course",
        "tags": [
          "Courses"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Course"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/reviews": {
      "get": {
        "summary": "Get all reviews",
        "tags": [
          "Reviews"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "pagination": {
                      "type": "object"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Review"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a review",
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Review"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "summary": "Get all users",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "pagination": {
                      "type": "object"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create a user",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
...
}